refactor: break ctx into modules
