+++ crates/rpc/src/lib.rs
@@ -53,7 +53,7 @@ mod config;
 pub use config::{RpcServerGuard, ServeConfig};
 
 mod ctx;
-pub use ctx::{RpcCtx, RuRevmState};
+pub use ctx::{RpcCtx, RuRevmState, SignetCtx};
 
 mod eth;
 pub use eth::{CallErrorData, EthError, eth};
